requirements.txt
Flask>=3.0,<4.0
Pillow>=10.0,<11.0
edge-tts>=6.1,<7.0
matplotlib>=3.8,<4.0
google-api-python-client>=2.0,<3.0
google-auth-oauthlib>=1.2,<2.0
google-auth-httplib2>=0.2,<1.0
